Data Engineering/FastAPI/pyproject.toml
[project]
name = "fastapi"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "aiomysql>=0.3.2",
    "aiosqlite>=0.22.1",
    "cloudinary>=1.44.1",
    "cryptography>=46.0.6",
    "imagekitio[aiohttp]>=5.2.0",
    "passlib[bcrypt]>=1.7.4",
    "pydantic>=2.12.5",
    "pyjwt>=2.12.1",
    "python-dotenv>=1.2.2",
    "python-multipart>=0.0.22",
    "sqlalchemy>=2.0.48",
]
